Move `PUT` method for responses under its own dynamic route.
